fix: jsdom teardown
When the teardown runs (when threads are disabled), it breaks JSDOM
It turns out that the teardown function is never called when running with threads enabled,
so this issue went unnoticed for most developers.

Fixes #835

The PR can be tested by running:

```
cd examples/vue
pnpm run test -- --no-threads
```

This fails on `main` but no longer fails with this commit.